fix(geometry): a zero-area triangle produced NaN normals through the CSG clipper
Last blocker on the freeze. CI's Moonshot lane failed at E2c because the F5
domain hardening on this branch REJECTS REAL KERNEL OUTPUT: `normals[96] must
be a finite number representable as f32 (got NaN)`. Two of the five documented
consumers of the frozen format -- g1 and g2-footprint -- were unrunnable, which
is not a state to freeze a 1.0.0 in.

F5 is right and was not touched. Every NaN bit pattern collapses to one quiet
NaN through setFloat64, so accepting NaN would give distinct payloads the same
node hash -- a second preimage a certificate would verify. Nothing was sanitised
in the demos either. The defect is upstream.

THE PRODUCER, localised by experiment rather than deduction:
csg/mod.rs Triangle::normal() called `.normalize()` on the edge cross product,
so a zero-area triangle yielded 0.0/0.0 = NaN. Its only production caller writes
that vector verbatim into Mesh::normals for every triangle clip_mesh emits, and
duplex reaches it through slice_mesh_into_layers -- the material-layer wall
slicing, which is why all 6 bad meshes are geometryClass 3. Pinned with a
sentinel build: setting the fallback to (0.6, 0.0, 0.8) put exactly that value
on exactly the 8 previously-NaN vertices of mesh #4465 and nowhere else.

Why it survived cleanup: clean_degenerate -> drop_thin_triangles rewrites only
`indices`. The degenerate triangle leaves the index buffer while its vertices
stay in positions/normals as orphans -- all 27 NaN vertices are unreferenced.

Convention: degenerate -> (0, 0, 1) in the kernel's Z-up frame, the same
fallback csg::normals::calculate_normals and weld_impl's average-normals path
already use, so consumers meet one convention rather than three. It has to be a
unit vector: (0,0,0) would recreate the division by zero in any shader or
exporter that re-normalizes. Non-degenerate triangles are bit-identical --
try_normalize(0.0) computes exactly what normalize() did for every non-zero
cross product.

WHAT MOVED, measured against a FRESH baseline wasm rather than the committed
pkg/ (which was stale from Jul 26 and gave 38,493 triangles against 39,334 for
a clean build -- branch drift, not this fix): 622 meshes both, 39,334 triangles
both, 81 NaN components in 6 meshes -> 0. Per-mesh digests differ on exactly
those 6 rows and no others. Five differ in normals only; mesh #4465 also drops
112 -> 108 vertices with the same 64 triangles, because NaN != NaN had been
preventing mesh_weld from merging coincident vertices.

NO FROZEN GOLDEN OR HASH VECTOR MOVED: assert-b35-golden matches byte-for-byte
(8931 bytes, all five acts ok) and demo-report.json is not dirty. g1 and g2
write no artifacts, so there is nothing to re-bless.

Triangle::normal's inline test module moved to csg/csg_tests.rs because the
module-size ratchet caps csg/mod.rs at exactly its previous 947 lines, so
nothing could be added without it; the file is now 788 and the allowlist is
ratcheted down to match. Two new tests: degenerate collapsed and collinear
triangles get +Z and are finite, and no mesh leaving clip_mesh carries a
non-finite normal.

Verified: g2-footprint exit 0 (falseConflictRate 0.075, missed 0, kill
criterion PASS), g1 exit 0 (correctness cross-check PASS), cargo test
-p ifc-lite-geometry 697 passed / 0 failed, -p ifc-lite-processing 131 passed,
module-size ratchet 4/4, assert-b35-golden exit 0, numeral gate exit 0.

Two hazards deliberately left, each needing its own before/after: the
drop_*_triangles paths do not compact vertices, so 345 orphans still ship for
duplex (clip_triangles_to_aabb in the same file does compact, and documents
why); and Plane::new has the identical unguarded normalize(), with no current
caller able to feed it a zero vector.
