+++ rust/geometry/src/csg/mod.rs
@@ -73,12 +73,43 @@ impl Triangle {
         Self { v0, v1, v2 }
     }
 
-    /// Calculate triangle normal
+    /// Calculate triangle normal.
+    ///
+    /// **Degenerate triangles get `+Z`, never NaN.** A zero-area (collapsed or
+    /// exactly collinear) triangle has a zero-length cross product, and the
+    /// plain `normalize()` this used to call is `v / |v|` — i.e. `0.0 / 0.0`,
+    /// which is NaN in every component. Those NaNs were written verbatim into
+    /// `Mesh::normals` by `add_triangle_to_mesh` (the only production caller of
+    /// this method, via `ClippingProcessor::clip_mesh`), and they SURVIVED the
+    /// mesh-hygiene pass: `clean_degenerate` / `drop_thin_triangles` rewrites
+    /// only `indices`, so the degenerate triangle's vertices stay in
+    /// `positions` / `normals` as ORPHANS carrying NaN. Six of duplex.ifc's
+    /// material-layer wall slices shipped 81 NaN normal components that way,
+    /// which the `@ifc-lite/provenance` node-hash domain check rightly rejects
+    /// (every NaN bit pattern collapses to one quiet NaN when serialized, so
+    /// accepting them would give distinct payloads the same hash).
+    ///
+    /// `+Z` is this crate's established convention for an undefined normal —
+    /// the same fallback `csg::normals::calculate_normals` and
+    /// `mesh::weld_impl`'s average-normals path already use — so a consumer
+    /// that meets one meets them all. It is stated in the KERNEL's own Z-up
+    /// frame, like every other normal this crate writes, so a viewer that
+    /// converts to Y-up reads it back as `+Y`; that is the conversion doing its
+    /// job, not a second convention. The value is arbitrary but must be a FIXED
+    /// unit vector: a zero normal would just re-create the division by zero in
+    /// any shader or exporter that re-normalizes.
+    ///
+    /// Non-degenerate triangles are unaffected, bit-for-bit: `try_normalize(0.0)`
+    /// returns `Some(v.unscale(|v|))` for every `|v| > 0`, which is exactly what
+    /// `normalize()` computed. The extra `is_finite` check covers the
+    /// astronomically-unlikely underflow case where `|v|` rounds to zero from
+    /// non-zero components (division would yield ±Inf, also out of domain).
     #[inline]
     pub fn normal(&self) -> Vector3<f64> {
-        let edge1 = self.v1 - self.v0;
-        let edge2 = self.v2 - self.v0;
-        edge1.cross(&edge2).normalize()
+        match self.cross_product().try_normalize(0.0) {
+            Some(n) if n.x.is_finite() && n.y.is_finite() && n.z.is_finite() => n,
+            _ => Vector3::new(0.0, 0.0, 1.0),
+        }
     }
 
     /// Calculate the cross product of edges, which is twice the area vector.
@@ -753,195 +784,5 @@ fn add_triangle_to_mesh(mesh: &mut Mesh, triangle: &Triangle) {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// Build a box mesh from AABB min/max bounds (12 triangles, 2 per face).
-    /// Test-only fixture builder for `subtract_mesh_many_chunks_match_sequential`
-    /// below; production code has no AABB-box-to-mesh path (D10 dead-code sweep
-    /// deleted `subtract_box`/`aabb_to_mesh`, whose only callers were tests).
-    fn aabb_to_mesh(min: Point3<f64>, max: Point3<f64>) -> Mesh {
-        let mut mesh = Mesh::with_capacity(8, 36);
-
-        let v0 = Point3::new(min.x, min.y, min.z);
-        let v1 = Point3::new(max.x, min.y, min.z);
-        let v2 = Point3::new(max.x, max.y, min.z);
-        let v3 = Point3::new(min.x, max.y, min.z);
-        let v4 = Point3::new(min.x, min.y, max.z);
-        let v5 = Point3::new(max.x, min.y, max.z);
-        let v6 = Point3::new(max.x, max.y, max.z);
-        let v7 = Point3::new(min.x, max.y, max.z);
-
-        add_triangle_to_mesh(&mut mesh, &Triangle::new(v0, v2, v1));
-        add_triangle_to_mesh(&mut mesh, &Triangle::new(v0, v3, v2));
-        add_triangle_to_mesh(&mut mesh, &Triangle::new(v4, v5, v6));
-        add_triangle_to_mesh(&mut mesh, &Triangle::new(v4, v6, v7));
-        add_triangle_to_mesh(&mut mesh, &Triangle::new(v0, v4, v7));
-        add_triangle_to_mesh(&mut mesh, &Triangle::new(v0, v7, v3));
-        add_triangle_to_mesh(&mut mesh, &Triangle::new(v1, v2, v6));
-        add_triangle_to_mesh(&mut mesh, &Triangle::new(v1, v6, v5));
-        add_triangle_to_mesh(&mut mesh, &Triangle::new(v0, v1, v5));
-        add_triangle_to_mesh(&mut mesh, &Triangle::new(v0, v5, v4));
-        add_triangle_to_mesh(&mut mesh, &Triangle::new(v3, v7, v6));
-        add_triangle_to_mesh(&mut mesh, &Triangle::new(v3, v6, v2));
-
-        mesh
-    }
-
-    /// More cutters than MAX_CUTTERS_PER_ARRANGEMENT force the chunked path in
-    /// `subtract_mesh_many`; the result must match the sequential subtract chain.
-    /// Set difference is order-independent (`host - {all}` equals
-    /// `host - {chunk1} - {chunk2} - ...`), so chunking is solid-equivalent. Guards
-    /// the chunk boundary (the perf fix for the 86 MB model that stalled the
-    /// geometry stream on a ~90-opening host packed into one arrangement).
-    #[test]
-    fn subtract_mesh_many_chunks_match_sequential() {
-        fn vol(m: &Mesh) -> f64 {
-            let p = |i: u32| {
-                let k = i as usize * 3;
-                [
-                    m.positions[k] as f64,
-                    m.positions[k + 1] as f64,
-                    m.positions[k + 2] as f64,
-                ]
-            };
-            let mut v = 0.0;
-            for t in m.indices.chunks_exact(3) {
-                let (a, b, c) = (p(t[0]), p(t[1]), p(t[2]));
-                v += a[0] * (b[1] * c[2] - c[1] * b[2])
-                    - a[1] * (b[0] * c[2] - c[0] * b[2])
-                    + a[2] * (b[0] * c[1] - c[0] * b[1]);
-            }
-            (v / 6.0).abs()
-        }
-        let csg = ClippingProcessor::new();
-        // Long wall + 20 disjoint through-openings (>16 ⇒ 2 chunks at the cap).
-        let wall = aabb_to_mesh(Point3::new(0., 0., 0.), Point3::new(40., 3., 0.2));
-        let cutters: Vec<Mesh> = (0..20)
-            .map(|i| {
-                let x = 1.0 + i as f64 * 2.0; // 2 m spacing ⇒ pairwise disjoint
-                aabb_to_mesh(Point3::new(x, 1., -0.5), Point3::new(x + 1.0, 2., 0.7))
-            })
-            .collect();
-        let refs: Vec<&Mesh> = cutters.iter().collect();
-        let batched = csg
-            .subtract_mesh_many(&wall, &refs)
-            .expect("chunked subtract must conform");
-        let mut seq = wall.clone();
-        for c in &cutters {
-            seq = csg.subtract_mesh(&seq, c).expect("sequential subtract");
-        }
-        let (vb, vs) = (vol(&batched), vol(&seq));
-        assert!(
-            (vb - vs).abs() < 1e-4,
-            "chunked volume {vb} != sequential {vs} on 20 disjoint cutters"
-        );
-        // Sanity: ~20 holes (~0.2 m³ each) actually removed from the ~24 m³ wall.
-        assert!(
-            vb < vol(&wall) - 3.0,
-            "expected ~20 holes removed; wall {} -> {vb}",
-            vol(&wall)
-        );
-    }
-
-    #[test]
-    fn test_plane_signed_distance() {
-        let plane = Plane::new(Point3::new(0.0, 0.0, 0.0), Vector3::new(0.0, 0.0, 1.0));
-
-        assert_eq!(plane.signed_distance(&Point3::new(0.0, 0.0, 5.0)), 5.0);
-        assert_eq!(plane.signed_distance(&Point3::new(0.0, 0.0, -5.0)), -5.0);
-        assert_eq!(plane.signed_distance(&Point3::new(5.0, 5.0, 0.0)), 0.0);
-    }
-
-    #[test]
-    fn test_clip_triangle_all_front() {
-        let processor = ClippingProcessor::new();
-        let triangle = Triangle::new(
-            Point3::new(0.0, 0.0, 1.0),
-            Point3::new(1.0, 0.0, 1.0),
-            Point3::new(0.5, 1.0, 1.0),
-        );
-        let plane = Plane::new(Point3::new(0.0, 0.0, 0.0), Vector3::new(0.0, 0.0, 1.0));
-
-        match processor.clip_triangle(&triangle, &plane) {
-            ClipResult::AllFront(_) => {}
-            _ => panic!("Expected AllFront"),
-        }
-    }
-
-    #[test]
-    fn test_clip_triangle_all_behind() {
-        let processor = ClippingProcessor::new();
-        let triangle = Triangle::new(
-            Point3::new(0.0, 0.0, -1.0),
-            Point3::new(1.0, 0.0, -1.0),
-            Point3::new(0.5, 1.0, -1.0),
-        );
-        let plane = Plane::new(Point3::new(0.0, 0.0, 0.0), Vector3::new(0.0, 0.0, 1.0));
-
-        match processor.clip_triangle(&triangle, &plane) {
-            ClipResult::AllBehind => {}
-            _ => panic!("Expected AllBehind"),
-        }
-    }
-
-    #[test]
-    fn test_clip_triangle_split_one_front() {
-        let processor = ClippingProcessor::new();
-        let triangle = Triangle::new(
-            Point3::new(0.0, 0.0, 1.0),  // Front
-            Point3::new(1.0, 0.0, -1.0), // Behind
-            Point3::new(0.5, 1.0, -1.0), // Behind
-        );
-        let plane = Plane::new(Point3::new(0.0, 0.0, 0.0), Vector3::new(0.0, 0.0, 1.0));
-
-        match processor.clip_triangle(&triangle, &plane) {
-            ClipResult::Split(triangles) => {
-                assert_eq!(triangles.len(), 1);
-            }
-            _ => panic!("Expected Split"),
-        }
-    }
-
-    #[test]
-    fn test_clip_triangle_split_two_front() {
-        let processor = ClippingProcessor::new();
-        let triangle = Triangle::new(
-            Point3::new(0.0, 0.0, 1.0),  // Front
-            Point3::new(1.0, 0.0, 1.0),  // Front
-            Point3::new(0.5, 1.0, -1.0), // Behind
-        );
-        let plane = Plane::new(Point3::new(0.0, 0.0, 0.0), Vector3::new(0.0, 0.0, 1.0));
-
-        match processor.clip_triangle(&triangle, &plane) {
-            ClipResult::Split(triangles) => {
-                assert_eq!(triangles.len(), 2);
-            }
-            _ => panic!("Expected Split with 2 triangles"),
-        }
-    }
-
-    #[test]
-    fn test_triangle_normal() {
-        let triangle = Triangle::new(
-            Point3::new(0.0, 0.0, 0.0),
-            Point3::new(1.0, 0.0, 0.0),
-            Point3::new(0.0, 1.0, 0.0),
-        );
-
-        let normal = triangle.normal();
-        assert!((normal.z - 1.0).abs() < 1e-6);
-    }
-
-    #[test]
-    fn test_triangle_area() {
-        let triangle = Triangle::new(
-            Point3::new(0.0, 0.0, 0.0),
-            Point3::new(1.0, 0.0, 0.0),
-            Point3::new(0.0, 1.0, 0.0),
-        );
-
-        let area = triangle.area();
-        assert!((area - 0.5).abs() < 1e-6);
-    }
-}
+#[path = "csg_tests.rs"]
+mod csg_tests;
