+++ rust/processing/tests/module_size_allowlist.txt
@@ -35,7 +35,9 @@
 # flood that open segments would corrupt.
      1617 rust/geometry/src/cdt.rs
      956 rust/geometry/src/csg/consolidate.rs
-     947 rust/geometry/src/csg/mod.rs
+# -159: inline `mod tests` split out to csg/csg_tests.rs (ratchet-exempt) to make room for the
+# degenerate-triangle NaN-normal fix and its two regression tests; budget refreshed downward.
+     788 rust/geometry/src/csg/mod.rs
  # -73: #1806 watertight-extrude tests split out to extrusion_watertight_tests.rs (ratchet-exempt); budget refreshed downward.
 # +9: #1806 review — extrude_profile_watertight rejects < 3-vertex hole rings before
 # triangulation (triangulate_constrained drops them from the cap while create_side_walls
